relocate history code
