+++ packages/sanddance-explorer/src/dialogs/history.tsx
@@ -5,12 +5,27 @@
 
 import { base } from '../base';
 import { Group } from '../controls/group';
-import { HistoryItem } from '../explorer';
 import { KeyCodes } from '../keycodes';
 import { strings } from '../language';
 import { FluentUITypes } from '@msrvida/fluentui-react-cdn-typings';
-import { util } from '@msrvida/sanddance-react';
+import { SandDance, util } from '@msrvida/sanddance-react';
 import { Explorer_Class } from '../explorer';
+
+export interface HistoricInsight extends SandDance.specs.Insight {
+    rebaseFilter?: boolean;
+    setup?: SandDance.types.Setup;
+}
+
+export interface HistoryAction {
+    insert?: boolean;
+    omit?: boolean;
+    label: string;
+}
+
+export interface HistoryItem {
+    label: string;
+    historicInsight: Partial<HistoricInsight>;
+}
 
 export interface Props {
     disabled?: boolean;
@@ -55,4 +70,44 @@ export function History(props: Props) {
             </ol>
         </Group>
     );
+}
+
+export function getPureInsight(historicInsight: HistoricInsight) {
+    const { colorBin, columns, directColor, facetStyle, filter, hideAxes, hideLegend, scheme, signalValues, size, totalStyle, transform, chart, view } = historicInsight;
+    const insight: SandDance.specs.Insight = {
+        colorBin,
+        columns,
+        directColor,
+        facetStyle,
+        filter,
+        hideAxes,
+        hideLegend,
+        scheme,
+        signalValues,
+        size,
+        totalStyle,
+        transform,
+        chart,
+        view,
+    };
+    return insight;
+}
+
+export function replay(historyItems: HistoryItem[], index: number): Partial<HistoricInsight> {
+    let filter: SandDance.searchExpression.Search = null;
+    let historicInsight: Partial<HistoricInsight> = {};
+    for (let i = 0; i < index + 1; i++) {
+        const historyItem = historyItems[i];
+        if (historyItem) {
+            if (historyItem.historicInsight.filter === null) {
+                filter = null;
+            } else if (historyItem.historicInsight.rebaseFilter) {
+                filter = historyItem.historicInsight.filter;
+            } else if (historyItem.historicInsight.filter) {
+                filter = SandDance.searchExpression.narrow(filter, historyItem.historicInsight.filter);
+            }
+            historicInsight = { ...historicInsight, ...historyItem.historicInsight };
+        }
+    }
+    return { ...historicInsight, filter };
 }
